FROMLIST: video: add HDMI state notifier support
Add support for HDMI hotplug and EDID notifiers, which is used to convey
information from HDMI drivers to their CEC and audio counterparts.

Based on an earlier version from Russell King:

https://patchwork.kernel.org/patch/9277043/

The hdmi_notifier is a reference counted object containing the HDMI state
of an HDMI device.

When a new notifier is registered the current state will be reported to
that notifier at registration time.

RFCv2:
https://patchwork.kernel.org/patch/9427683/
https://git.linuxtv.org/hverkuil/media_tree.git/log/?h=cec-notifiers

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>